backend/requirements.txt
faker
neo4j
python-dotenv
fastapi
uvicorn
pydantic
requests
numpy
pandas
concurrently
joblib